Redo Countdown Creation
Everything is now defined in a 'countdown object', which contains
everything a countdown needs to know. Has an array of times, so the
countdown can get straight on to the next one.

I suppose this has come about due to wanting to create a generic timer
system that could be used for other purposes. Perhaps I'll set up
another repo at some time which will be just a generic timer that can be
used for anything.