Credit artists for their work
I've been meaning to do this for a while. I wasn't sure how I wanted to
do it, but it needed to be done.

I feel there's too much going on in the centre of the screen. Granted
that's where the attention should be, but it just feels too busy.