Remove setInterval, use requestAnimationFrame
This way the countdown only updates when the browser's renderer is about
to run again. This prevents the event stack getting clogged up on slower
systems (such as mobile devices).

Now can use startCountdowns() and stopCountdowns() from dev tools
console, in case that's needed. Due to the way the countdown works, when
it's resumed it will jump to where it would have been anyway.